requirements.txt
Flask
opencv-python-headless
numpy
pyzbar
python-dotenv
